Add: Sale service and sale repository refactored
