+++ src/main/java/com/optivision/webapp/sale/exception/InvalidSaleDataException.java
@@ -1,4 +1,4 @@
-package com.optivision.webapp.venta.exception;
+package com.optivision.webapp.sale.exception;
 
 import lombok.NoArgsConstructor;
